Move `h.views.api_config` -> `h.views.api.config`
- Update dependencies in other modules, tests
